Rename Rust module lib->utils, since 'lib' becomes a reserved module name in future Rust versions
