+++ src/audit_logs.rs
@@ -1,7 +1,7 @@
 use crate::cli::{FORMAT_OPT, LIST_SUBCMD};
 use crate::database::{AuditLogs, Environments, OpenApiConfig, Parameters, Projects, Users};
-use crate::lib::{error_message, help_message, parse_datetime, warn_missing_subcommand};
 use crate::table::Table;
+use crate::utils::{error_message, help_message, parse_datetime, warn_missing_subcommand};
 use clap::ArgMatches;
 use color_eyre::eyre::Result;
 use indoc::printdoc;
