+++ src/login.rs
@@ -1,6 +1,6 @@
 use crate::cli::{binary_name, CONFIRM_FLAG};
 use crate::config::Config;
-use crate::lib::{get_api_access_url, user_confirm, warning_message, API_KEY_PAGE, SEPARATOR};
+use crate::utils::{get_api_access_url, user_confirm, warning_message, API_KEY_PAGE, SEPARATOR};
 use clap::ArgMatches;
 use color_eyre::eyre::Result;
 use indoc::printdoc;
