+++ src/main.rs
@@ -13,7 +13,6 @@ mod groups;
 mod import;
 mod installation;
 mod integrations;
-mod lib;
 mod login;
 mod logout;
 mod parameters;
@@ -25,6 +24,7 @@ mod table;
 mod templates;
 mod types;
 mod users;
+mod utils;
 mod versions;
 
 use crate::actions::process_actions_command;
@@ -40,7 +40,6 @@ use crate::groups::process_groups_command;
 use crate::import::process_import_command;
 use crate::installation::{binary_version, get_latest_version, install_latest_version};
 use crate::integrations::process_integrations_command;
-use crate::lib::{error_message, help_message, warning_message};
 use crate::login::process_login_command;
 use crate::logout::process_logout_command;
 use crate::parameters::process_parameters_command;
@@ -50,6 +49,7 @@ use crate::schema::process_schema_command;
 use crate::templates::process_templates_command;
 use crate::types::process_parameter_type_command;
 use crate::users::process_users_command;
+use crate::utils::{error_message, help_message, warning_message};
 use crate::versions::process_version_command;
 use chrono::{Datelike, NaiveDate, Utc};
 use clap::ArgMatches;
